+++ MineBomber/src/com/me/Map/TilesInfo.java
@@ -7,7 +7,7 @@ import com.badlogic.gdx.graphics.g2d.TextureRegion;
  * Created by alekseev on 21.03.2014.
  */
 public class TilesInfo {
-    public float mLife;
+    public int mLife;
     public int mId;
     public int mNextid;
     public TextureRegion mTexRegion;
@@ -18,7 +18,7 @@ public class TilesInfo {
 
     public final static int TYPE_GROUND=0;
 
-     public TilesInfo(int id,int nextid,int type,float life,TextureRegion textureRegion)
+     public TilesInfo(int id,int nextid,int type,int life,TextureRegion textureRegion)
      {
          mLife=life;
          mId=id;
